+++ meta_package_manager/managers/apm.py
@@ -22,7 +22,7 @@ from boltons.cacheutils import cachedproperty
 
 from ..base import PackageManager
 from ..platform import LINUX, MACOS, WINDOWS
-from ..version import parse_version
+from ..version import TokenizedString, parse_version
 
 
 class APM(PackageManager):
@@ -153,7 +153,7 @@ class APM(PackageManager):
 
         return installed
 
-    def search(self, query):
+    def search(self, query, extended, exact):
         """ Fetch matching packages from ``apm search`` output.
 
         Raw CLI output samples:
@@ -237,17 +237,34 @@ class APM(PackageManager):
         """
         matches = {}
 
+        search_args = []
+        if not extended:
+            search_args.append('--no-description')
+
         output = self.run([self.cli_path] + self.global_args + [
-            'search', query, '--json'])
+            'search', '--json'] + search_args + [query])
 
         if output:
             for package in json.loads(output):
                 package_id = package['name']
+
+                # Exclude packages not featuring the search query in their ID
+                # or name.
+                if not extended:
+                    query_parts = set(map(str, TokenizedString(query)))
+                    pkg_parts = set(map(str, TokenizedString(package_id)))
+                    if not query_parts.issubset(pkg_parts):
+                        continue
+
+                # Filters out fuzzy matches, only keep stricly matching
+                # packages.
+                if exact and query != package_id:
+                    continue
+
                 matches[package_id] = {
                     'id': package_id,
                     'name': package_id,
-                    'latest_version': parse_version(package['version']),
-                    'exact': self.exact_match(query, package_id)}
+                    'latest_version': parse_version(package['version'])}
 
         return matches